hit class split to Track & Hit class, hit removed
